refactor(weekly): page.tsx 8517行からモーダル20個を分離
- src/app/(main)/menus/weekly/_components/modals/*.tsx 配下に
  AddMealModal, FridgeModal, ShoppingModal 等 20モーダルを切り出し
- ProgressTodoCard も _components/ に分離
- page.tsx は -3050 / +467 で大幅縮小
- state 受け渡しと handler は page.tsx 側に維持
- 既存の testID / aria-label は保持
- 関連: docs/refactor/2026-05-08-tech-debt-elimination.md
